Added IdentifyTextResultTransformers+Tables.swift and IdentifyTextResultTransformers.swift
